+++ FilmApps/app/src/main/java/com/example/filmapps/Screens.kt
@@ -1,10 +1,9 @@
 package com.example.filmapps
 
-import android.os.Bundle
 import com.example.filmapps.feature.regAndAuth.authorization.presentation.ui.AuthorizationFragment
 import com.example.filmapps.feature.regAndAuth.registrations.presentation.ui.RegistrationFragment
-import com.example.filmapps.presentation.model.Character
-import com.example.filmapps.presentation.ui.CharacterListFragment
+import com.example.filmapps.feature.characterList.presentation.model.Character
+import com.example.filmapps.feature.characterList.presentation.ui.CharacterListFragment
 import com.example.filmapps.presentation.ui.DetailsFragment
 import com.github.terrakok.cicerone.androidx.FragmentScreen
