--- FilmApps/app/src/main/java/com/example/filmapps/data/repository/RetrofitRepository.kt
@@ -1,7 +1,0 @@
-package com.example.filmapps.data.repository
-
-import com.example.filmapps.presentation.model.CharacterListResponce
-
-interface RetrofitRepository {
-    suspend fun loadList(): CharacterListResponce
-}
