--- FilmApps/app/src/main/java/com/example/filmapps/domain/useCase/GetListCharacterUseCase.kt
@@ -1,7 +1,0 @@
-package com.example.filmapps.domain.useCase
-
-import com.example.filmapps.presentation.model.CharacterListResponce
-
-interface GetListCharacterUseCase {
-    suspend fun execute(): CharacterListResponce
-}
