+++ FilmApps/app/src/main/java/com/example/filmapps/feature/characterList/data/repository/RetrofitRepository.kt
@@ -1,0 +1,7 @@
+package com.example.filmapps.feature.characterList.data.repository
+
+import com.example.filmapps.feature.characterList.presentation.model.CharacterListResponce
+
+interface RetrofitRepository {
+    suspend fun loadList(): CharacterListResponce
+}
