+++ FilmApps/app/src/main/java/com/example/filmapps/feature/characterList/data/repository/RetrofitRepositoryImpl.kt
@@ -1,12 +1,12 @@
-package com.example.filmapps.data.repository
+package com.example.filmapps.feature.characterList.data.repository
 
-import com.example.filmapps.data.retrofit.GetPage
-import com.example.filmapps.presentation.model.CharacterListResponce
+import com.example.filmapps.feature.characterList.data.retrofit.GetPageApi
+import com.example.filmapps.feature.characterList.presentation.model.CharacterListResponce
 import java.lang.Exception
 import javax.inject.Inject
 
 class RetrofitRepositoryImpl @Inject constructor(
-    private val retrofit: GetPage
+    private val retrofit: GetPageApi
 ) : RetrofitRepository {
     var page = 0
     private val lastPage = 42
