+++ FilmApps/app/src/main/java/com/example/filmapps/feature/characterList/data/retrofit/GetPageApi.kt
@@ -1,0 +1,11 @@
+package com.example.filmapps.feature.characterList.data.retrofit
+
+import com.example.filmapps.feature.characterList.data.model.DataApi
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface GetPageApi {
+
+    @GET("character/")
+    suspend fun getPage(@Query("page") page: Int): DataApi
+}
