+++ FilmApps/app/src/main/java/com/example/filmapps/feature/characterList/di/scope/CharacterListScope.kt
@@ -1,4 +1,4 @@
-package com.example.filmapps.di.scope
+package com.example.filmapps.feature.characterList.di.scope
 
 import javax.inject.Scope
