+++ FilmApps/app/src/main/java/com/example/filmapps/feature/characterList/domain/useCase/GetListCharacterUseCase.kt
@@ -1,0 +1,7 @@
+package com.example.filmapps.feature.characterList.domain.useCase
+
+import com.example.filmapps.feature.characterList.presentation.model.CharacterListResponce
+
+interface GetListCharacterUseCase {
+    suspend fun execute(): CharacterListResponce
+}
