+++ FilmApps/app/src/main/java/com/example/filmapps/feature/characterList/domain/useCase/GetListCharacterUseCaseImpl.kt
@@ -1,7 +1,7 @@
-package com.example.filmapps.domain.useCase
+package com.example.filmapps.feature.characterList.domain.useCase
 
-import com.example.filmapps.data.repository.RetrofitRepository
-import com.example.filmapps.presentation.model.CharacterListResponce
+import com.example.filmapps.feature.characterList.data.repository.RetrofitRepository
+import com.example.filmapps.feature.characterList.presentation.model.CharacterListResponce
 import javax.inject.Inject
 
 class GetListCharacterUseCaseImpl @Inject constructor(
