+++ FilmApps/app/src/main/java/com/example/filmapps/feature/characterList/presentation/model/Character.kt
@@ -1,0 +1,3 @@
+package com.example.filmapps.feature.characterList.presentation.model
+
+data class Character(val id: String?, val img: String?)
