+++ FilmApps/app/src/main/java/com/example/filmapps/feature/characterList/presentation/model/CharacterListResponce.kt
@@ -1,6 +1,6 @@
-package com.example.filmapps.presentation.model
+package com.example.filmapps.feature.characterList.presentation.model
 
-import com.example.filmapps.data.model.DataApi
+import com.example.filmapps.feature.characterList.data.model.DataApi
 import java.lang.Exception
 
 sealed class CharacterListResponce {
