--- FilmApps/app/src/main/java/com/example/filmapps/presentation/model/Character.kt
@@ -1,3 +1,0 @@
-package com.example.filmapps.presentation.model
-
-data class Character(val id: String?, val img: String?)
