+++ FilmApps/app/src/main/res/layout/fragment_character_list_list.xml
@@ -13,7 +13,7 @@
         android:layout_height="match_parent"
         app:spanCount="3"
         app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
-        tools:context=".presentation.ui.CharacterListFragment"
+        tools:context=".feature.characterList.presentation.ui.CharacterListFragment"
         tools:listitem="@layout/fragment_character_list" />
 
     <ProgressBar
